backend/requirements.txt
asgiref
certifi
cffi
chardet
cryptography
defusedxml
Django
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
idna
oauthlib

pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
sqlparse
urllib3
